requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
requests
werkzeug
pymysql
python-dotenv
SpeechRecognition
pydub
google-generativeai
